refactor(payment):usePrice훅으로 가격변수 분리
